Implemented Color Sensing Algorithm
This was a result of looking at the Matlab plot, segments the color space into acceptable regions which are (mostly) exclusive to tape strips of a given color. Also implemented a filter which only flags a success if the sensor has been over the color strip for a certain number of cycles. I think this will be more effective than a running average filter because with the nature of the fluctuation over the white tape strip specifically, it's possible that we'll have some random false positives due to the average. Conceptually, this solution should be fairly robust, we just need to tune the parameters. Next step is to gather a ton of data with the sensor on the actual chassis.